Initial Project
This is written in a combination of C# and HTML. We are developing this using Blazor, a new framework for web development put out by Microsoft. It is intended to get rid of needing custom javascript for styling, we can use open source packages instead. The open source package that I have already started using is Radzen. This is C#, an even easier version of programming C++. It kinda feels like programming in both Python and C++, its nice and easy to pick up since we all have done C and C++.