Many changes
I removed the history page and integrated it into the checkin page. I fixed the history queries so that they can be sorted and displayed in different quantities. I tied up a bunch of edge cases. I added a location and # of entries dropdown to thhe history section. I changed the input to take the full output of the vtap, including the card serial. I added functions to parse the new input scheme. I cleaned up how user data is maintained in the program. I adjusted the add user and edit user pages to a new format to account for a change in the fields of the database. I modified some of elizabeths changes to the user management page to have the card icon display in red instead of the entire row. I added an add user button to the user management page.